requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
diffusers==0.23.1
transformers==4.35.2
accelerate==0.24.1
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.2
